added uin8 as default sequence dtype for deployment scripts

+++ tensorflow2.0_compatibility_version/run_deploy_shape_combination_deepCregr.py
@@ -82,7 +82,7 @@ if FLAGS.slize != 'all':
 def get_hot_coded_seq(sequence, use_soft=False):
     """Convert a 4 base letter sequence to 4-row x-cols hot coded sequence"""
     # initialise empty
-    hotsequence = np.zeros((len(sequence),4))
+    hotsequence = np.zeros((len(sequence),4), dtype = 'uint8')
 
     # transform to uppercase if using softmasked sequences
     if use_soft:
